Use enums instead of uint where appropriate.

+++ tests/rccspec.cpp
@@ -18,39 +18,39 @@ void TestRCCSpec::testRCCSpec()
 
     spec = RequestableChannelClassSpec::textChat();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_TEXT);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeContact));
+    QCOMPARE(spec.targetHandleType(), HandleTypeContact);
     QCOMPARE(spec.fixedProperties().size(), 2);
     QCOMPARE(spec.allowedProperties().isEmpty(), true);
 
     spec = RequestableChannelClassSpec::textChatroom();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_TEXT);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeRoom));
+    QCOMPARE(spec.targetHandleType(), HandleTypeRoom);
     QCOMPARE(spec.fixedProperties().size(), 2);
     QCOMPARE(spec.allowedProperties().isEmpty(), true);
 
     spec = RequestableChannelClassSpec::streamedMediaCall();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_STREAMED_MEDIA);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeContact));
+    QCOMPARE(spec.targetHandleType(), HandleTypeContact);
     QCOMPARE(spec.fixedProperties().size(), 2);
     QCOMPARE(spec.allowedProperties().isEmpty(), true);
 
     spec = RequestableChannelClassSpec::streamedMediaAudioCall();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_STREAMED_MEDIA);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeContact));
+    QCOMPARE(spec.targetHandleType(), HandleTypeContact);
     QCOMPARE(spec.fixedProperties().size(), 2);
     QCOMPARE(spec.allowedProperties().size(), 1);
     QCOMPARE(spec.allowsProperty(TP_QT4_IFACE_CHANNEL_TYPE_STREAMED_MEDIA + QLatin1String(".InitialAudio")), true);
 
     spec = RequestableChannelClassSpec::streamedMediaVideoCall();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_STREAMED_MEDIA);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeContact));
+    QCOMPARE(spec.targetHandleType(), HandleTypeContact);
     QCOMPARE(spec.fixedProperties().size(), 2);
     QCOMPARE(spec.allowedProperties().size(), 1);
     QCOMPARE(spec.allowsProperty(TP_QT4_IFACE_CHANNEL_TYPE_STREAMED_MEDIA + QLatin1String(".InitialVideo")), true);
 
     spec = RequestableChannelClassSpec::streamedMediaVideoCallWithAudio();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_STREAMED_MEDIA);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeContact));
+    QCOMPARE(spec.targetHandleType(), HandleTypeContact);
     QCOMPARE(spec.fixedProperties().size(), 2);
     QCOMPARE(spec.allowedProperties().size(), 2);
     QCOMPARE(spec.allowsProperty(TP_QT4_IFACE_CHANNEL_TYPE_STREAMED_MEDIA + QLatin1String(".InitialAudio")), true);
@@ -71,13 +71,13 @@ void TestRCCSpec::testRCCSpec()
 
     spec = RequestableChannelClassSpec::conferenceTextChatroom();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_TEXT);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeRoom));
+    QCOMPARE(spec.targetHandleType(), HandleTypeRoom);
     QCOMPARE(spec.allowedProperties().size(), 1);
     QCOMPARE(spec.allowsProperty(TP_QT4_IFACE_CHANNEL_INTERFACE_CONFERENCE + QLatin1String(".InitialChannels")), true);
 
     spec = RequestableChannelClassSpec::conferenceTextChatroomWithInvitees();
     QCOMPARE(spec.channelType(), TP_QT4_IFACE_CHANNEL_TYPE_TEXT);
-    QCOMPARE(spec.targetHandleType(), static_cast<uint>(HandleTypeRoom));
+    QCOMPARE(spec.targetHandleType(), HandleTypeRoom);
     QCOMPARE(spec.allowedProperties().size(), 2);
     QCOMPARE(spec.allowsProperty(TP_QT4_IFACE_CHANNEL_INTERFACE_CONFERENCE + QLatin1String(".InitialChannels")), true);
     QCOMPARE(spec.allowsProperty(TP_QT4_IFACE_CHANNEL_INTERFACE_CONFERENCE + QLatin1String(".InitialInviteeHandles")), true);
